+++ bin/node/src/main.rs
@@ -6,7 +6,7 @@ use commands::{init::init_config_files, start::start_node};
 use miden_node_block_producer::server::BlockProducer;
 use miden_node_rpc::server::Rpc;
 use miden_node_store::server::Store;
-use miden_node_utils::{config::load_config, version::LongVersion};
+use miden_node_utils::{config::load_config, logging::OpenTelemetry, version::LongVersion};
 
 mod commands;
 mod config;
@@ -37,6 +37,9 @@ pub enum Command {
 
         #[arg(short, long, value_name = "FILE", default_value = NODE_CONFIG_FILE_PATH)]
         config: PathBuf,
+
+        #[arg(long = "open-telemetry", default_value_t = false)]
+        open_telemetry: bool,
     },
 
     /// Generates a genesis file and associated account files based on a specified genesis input
@@ -82,12 +85,18 @@ pub enum StartCommand {
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    miden_node_utils::logging::setup_logging()?;
-
     let cli = Cli::parse();
 
+    // Open telemetry exporting is only valid for running the node.
+    let open_telemetry = if let Command::Start { open_telemetry: true, .. } = &cli.command {
+        OpenTelemetry::Enabled
+    } else {
+        OpenTelemetry::Disabled
+    };
+    miden_node_utils::logging::setup_tracing(open_telemetry)?;
+
     match &cli.command {
-        Command::Start { command, config } => match command {
+        Command::Start { command, config, .. } => match command {
             StartCommand::Node => {
                 let config = load_config(config).context("Loading configuration file")?;
                 start_node(config).await