+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 ## Unreleased
 
+### Fixes
+
+- Faucet webpage is missing `background.png` and `favicon.ico` (#672).
+
+### Enhancements
+
+- Add an optional open-telemetry trace exporter (#659, #690).
+- Support tracing across gRPC boundaries using remote tracing context (#669).
+- Instrument the block-producer's block building process (#676).
+
+### Changes
+
+- [BREAKING] Updated minimum Rust version to 1.84.
+- [BREAKING] `Endpoint` configuration simplified to a single string (#654).
+
+### Enhancements
+
+- Prove transaction batches using Rust batch prover reference implementation (#659).
+
 ## v0.7.2 (2025-01-29)
 
 ### Fixes